Explain parameters of lisa & Move backend codes to src dir
