+++ python/llm/example/GPU/LLM-Finetuning/LISA/README.md
@@ -34,6 +34,17 @@ python ./lisa_finetuning.py \
     --lisa_activated_layers 1 \
     --lisa_interval_steps 20
 ```
+Optional parameters for `lisa_finetuning.py`:
+
+**--repo-id-or-model-path** : default to `meta-llama/Llama-2-7b-hf`, and you can also specify your local model path.
+
+**--data-path** : default to `yahma/alpaca-cleaned`, and you can also specify your local datal path, while note that changing to the other datasets will introduce code modification effort for yourself.
+
+**--output-dir** : default to `./ipex-llm-lisa-alpaca` to save fine-tuned model, and you can change if needed.
+
+**--lisa_activated_layers** :  the number of self-attention layers randomly selected to activate.
+
+**lisa_interval_steps** : the number of interval steps to switch active layers.
 
 ### 3. Sample Output
 ```log
